First attempt of adding a recipe subpage
Added a subpage where one can filter for recipes. Search bar works for now and there is a first instance of a checkbox filter. Next goal is to let recipe suggestions be shown on the search page where on can click directly on the link to get to the recipe